Refactor FastSync actor into smaller classes
